backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
bcrypt==4.3.0
boto3==1.39.7
botocore==1.39.7
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
discord.py==2.5.2
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
jose==1.0.0
motor==3.7.0
multidict==6.6.0
packaging==25.0
passlib==1.7.4
platformdirs==4.3.8
propcache==0.3.2
psutil==5.9.6
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
s3transfer==0.13.0
s8==0.1.6
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zipp==3.20.0
